src/requirements.txt
grpcio==1.42.0
protobuf==3.19.4 
# PyJWT==2.4.0
# pytest==7.1.2
# pyyaml==6.0
cryptography==2.5
pymongo==3.12.0